Commit Helper and Config dependency
This initial commit serves the incremental approach to making the COD Services API fully available on GitHub. The platform is WebAPI in .Net and additional system libraries will be needed to run the code, which may be adapted to other languages as well. A list of libraries used in dependencies will be provided in the project wiki. This is an effort to satisfy requests for enhanced developer support of COD Services provided through UN OCHA coordination.