Run Python P-Coding script from Excel using xlwings
This is an instruction file to run python scripts from Excel using xlwings library. The Location_Coordinates.xlsx file is used in this example, with the Location_Coordinates.py script. This example will create an output.xlsx spreadsheet file with the appropriate p-code to the geographic coordinates listed in the input spreadsheet columns.