Bulk P-Code in Excel Version 2 document
This version 2 will provide for the upload of the original spreadsheet and results with the additional COD services  columns added based on the row lat/long. This differs from the original version, which will make a call to the API inside Excel and is slower and not useful with larger datasets and requires external feature support to Excel.